pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "condottiere"
version = "0.1.0"
description = "Mercenary Den Monitoring & Alerting Service for EVE Online"
requires-python = ">=3.12"

dependencies = [
  "fastapi",
  "uvicorn",
  "python-multipart",
  "python-dotenv",
  "itsdangerous",
  "sqlalchemy",
  "alembic",
  "cryptography",
  "httpx",
  "pydantic",
]

[project.optional-dependencies]
prod = [
  "psycopg[binary]>=3.1",
]

[tool.setuptools.packages.find]
include = ["app*"]
